refactor: expand all when torch is install
