+++ docarray/typing/__init__.py
@@ -1,13 +1,5 @@
 from docarray.typing.id import ID
 from docarray.typing.tensor import NdArray, Tensor
-
-try:
-    import torch  # noqa: F401
-except ImportError:
-    pass
-else:
-    from docarray.typing.tensor import TorchEmbedding, TorchTensor  # noqa: F401
-
 from docarray.typing.tensor.embedding import Embedding
 from docarray.typing.url import AnyUrl, ImageUrl, TextUrl
