+++ docarray/typing/tensor/__init__.py
@@ -2,15 +2,6 @@ from docarray.typing.tensor.embedding import Embedding, NdArrayEmbedding
 from docarray.typing.tensor.ndarray import NdArray
 from docarray.typing.tensor.tensor import Tensor
 
-try:
-    import torch  # noqa: F401
-except ImportError:
-    pass
-else:
-    from docarray.typing.tensor.embedding import TorchEmbedding  # noqa: F401
-    from docarray.typing.tensor.torch_tensor import TorchTensor  # noqa: F401
-
-
 __all__ = [
     'NdArray',
     'Tensor',
