+++ docarray/typing/tensor/embedding/__init__.py
@@ -1,12 +1,13 @@
+from docarray.typing.tensor.embedding.embedding import Embedding
+from docarray.typing.tensor.embedding.ndarray import NdArrayEmbedding
+
+__all__ = ['NdArrayEmbedding', 'Embedding']
+
 try:
     import torch  # noqa: F401
 except ImportError:
     pass
 else:
-    from docarray.typing.tensor.embedding.embedding_torch import TorchEmbedding  # type: ignore # noqa: [F401,E510]
+    from docarray.typing.tensor.embedding.torch import TorchEmbedding  # noqa F401
 
-
-from docarray.typing.tensor.embedding.embedding import Embedding
-from docarray.typing.tensor.embedding.embedding_ndarray import NdArrayEmbedding
-
-__all__ = ['NdArrayEmbedding', 'Embedding']
+    __all__.append('TorchEmbedding')
