--- docarray/typing/tensor/embedding/embedding_ndarray.py
@@ -1,6 +1,0 @@
-from docarray.typing.tensor.embedding.embedding_mixin import EmbeddingMixin
-from docarray.typing.tensor.ndarray import NdArray
-
-
-class NdArrayEmbedding(NdArray, EmbeddingMixin):
-    alternative_type = NdArray
